+++ .storybook/main.js
@@ -18,6 +18,7 @@ module.exports = {
         },
       },
     },
+    '@storybook/addon-interactions',
   ],
   framework: '@storybook/react',
   features: {