+++ apps/remix/app/data.server.ts
@@ -19,6 +19,14 @@ export async function getComponents() {
       name: 'Lozenge',
     },
     {
+      to: 'components/modal',
+      name: 'Modal',
+    },
+    {
+      to: 'components/popover',
+      name: 'Popover',
+    },
+    {
       to: 'components/progress',
       name: 'Progress',
     },