+++ apps/remix/app/root.tsx
@@ -19,7 +19,10 @@ import lozengeStyles from '@launchpad-ui/lozenge/styles.css';
 import alertStyles from '@launchpad-ui/alert/styles/Alert.css';
 import toggleStyles from '@launchpad-ui/toggle/styles.css';
 import progressStyles from '@launchpad-ui/progress/styles.css';
-import TabListStyles from '@launchpad-ui/tab-list/styles.css';
+import tabListStyles from '@launchpad-ui/tab-list/styles.css';
+import modalStyles from '@launchpad-ui/modal/styles/Modal.css';
+import promptStyles from '@launchpad-ui/modal/styles/Prompt.css';
+import popoverStyles from '@launchpad-ui/popover/styles.css';
 
 export const links: LinksFunction = () => {
   return [
@@ -31,7 +34,10 @@ export const links: LinksFunction = () => {
     { rel: 'stylesheet', href: alertStyles },
     { rel: 'stylesheet', href: toggleStyles },
     { rel: 'stylesheet', href: progressStyles },
-    { rel: 'stylesheet', href: TabListStyles },
+    { rel: 'stylesheet', href: tabListStyles },
+    { rel: 'stylesheet', href: modalStyles },
+    { rel: 'stylesheet', href: promptStyles },
+    { rel: 'stylesheet', href: popoverStyles },
   ];
 };