+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @launchpad-ui/tokens
 
+## 0.1.3
+
+### Patch Changes
+
+- [#82](https://github.com/launchdarkly/launchpad-ui/pull/82) [`45e786a`](https://github.com/launchdarkly/launchpad-ui/commit/45e786a9972406fbf0f93f73aeeaa81d1fded4f8) Thanks [@Niznikr](https://github.com/Niznikr)! - Add overlay package:
+
+  - Add overlay code and tests
+  - Add portal to modal package
+  - Add custom media queries to tokens package
+
 ## 0.1.2
 
 ### Patch Changes